Develop

+++ src/main/java/ca/tunestumbler/api/ui/controller/UserController.java
@@ -3,9 +3,14 @@ package ca.tunestumbler.api.ui.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,14 +21,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import ca.tunestumbler.api.exceptions.UserServiceException;
+import com.google.common.base.Strings;
+
+import ca.tunestumbler.api.exceptions.InvalidBodyException;
+import ca.tunestumbler.api.exceptions.MissingPathParametersException;
 import ca.tunestumbler.api.service.UserService;
 import ca.tunestumbler.api.shared.dto.UserDTO;
 import ca.tunestumbler.api.ui.model.request.UserDetailsRequestModel;
 import ca.tunestumbler.api.ui.model.response.ErrorMessages;
-import ca.tunestumbler.api.ui.model.response.OperationStatusModel;
-import ca.tunestumbler.api.ui.model.response.RequestOperationName;
-import ca.tunestumbler.api.ui.model.response.RequestOperationStatus;
+import ca.tunestumbler.api.ui.model.response.ErrorPrefixes;
 import ca.tunestumbler.api.ui.model.response.UserDetailsResponseModel;
 
 @RestController
@@ -35,6 +41,11 @@ public class UserController {
 
 	@GetMapping(path = "/{userId}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public UserDetailsResponseModel getUser(@PathVariable String userId) {
+		if (Strings.isNullOrEmpty(userId)) {
+			throw new MissingPathParametersException(ErrorPrefixes.USER_CONTROLLER.getErrorPrefix()
+					+ ErrorMessages.MISSING_REQUIRED_PATH_FIELD.getErrorMessage());
+		}
+
 		UserDetailsResponseModel existingUserResponse = new UserDetailsResponseModel();
 
 		UserDTO userDTO = userService.getUserByUserId(userId);
@@ -44,7 +55,13 @@ public class UserController {
 	}
 
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public UserDetailsResponseModel createUser(@RequestBody UserDetailsRequestModel userDetails) throws Exception {
+	public ResponseEntity<?> createUser(@Valid @RequestBody UserDetailsRequestModel userDetails,
+			BindingResult bindingResult) {
+		if (bindingResult.hasErrors()) {
+			throw new InvalidBodyException(ErrorPrefixes.USER_CONTROLLER.getErrorPrefix()
+					+ ErrorMessages.INVALID_BODY.getErrorMessage());
+		}
+
 		UserDetailsResponseModel newUserResponse = new UserDetailsResponseModel();
 
 		UserDTO userDTO = new UserDTO();
@@ -53,12 +70,22 @@ public class UserController {
 		UserDTO createdUser = userService.createUser(userDTO);
 		BeanUtils.copyProperties(createdUser, newUserResponse);
 
-		return newUserResponse;
+		return new ResponseEntity<>(newUserResponse, HttpStatus.CREATED);
 	}
 
 	@PutMapping(path = "/{userId}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public UserDetailsResponseModel updateUser(@PathVariable String userId,
-			@RequestBody UserDetailsRequestModel userDetails) {
+			@Valid @RequestBody UserDetailsRequestModel userDetails, BindingResult bindingResult) {
+		if (Strings.isNullOrEmpty(userId)) {
+			throw new MissingPathParametersException(ErrorPrefixes.USER_CONTROLLER.getErrorPrefix()
+					+ ErrorMessages.MISSING_REQUIRED_PATH_FIELD.getErrorMessage());
+		}
+
+		if (bindingResult.hasErrors()) {
+			throw new InvalidBodyException(ErrorPrefixes.USER_CONTROLLER.getErrorPrefix()
+					+ ErrorMessages.INVALID_BODY.getErrorMessage());
+		}
+
 		UserDetailsResponseModel updatedUserResponse = new UserDetailsResponseModel();
 
 		UserDTO userDTO = new UserDTO();
@@ -71,14 +98,15 @@ public class UserController {
 	}
 
 	@DeleteMapping(path = "/{userId}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public OperationStatusModel deleteUser(@PathVariable String userId) {
-		OperationStatusModel deletedUserResponse = new OperationStatusModel();
-		deletedUserResponse.setOperationName(RequestOperationName.DELETE.name());
+	public ResponseEntity<?> deleteUser(@PathVariable String userId) {
+		if (Strings.isNullOrEmpty(userId)) {
+			throw new MissingPathParametersException(ErrorPrefixes.USER_CONTROLLER.getErrorPrefix()
+					+ ErrorMessages.MISSING_REQUIRED_PATH_FIELD.getErrorMessage());
+		}
 
 		userService.deleteUser(userId);
-		deletedUserResponse.setOperationResult(RequestOperationStatus.SUCCESS.name());
 
-		return deletedUserResponse;
+		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 
 	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
